requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.9.2
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.2
boto3==1.41.2
botocore==1.41.2
brotli==1.2.0
CacheControl==0.14.4
cbor2==5.7.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.3
cssselect2==0.8.0
curl_cffi==0.14.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.7
django-browser-reload==1.21.0
django-fernet-fields-v2==0.9
django-formtools==2.5.1
django-otp==1.6.2
django-phonenumber-field==8.3.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-storages==1.14.6
django-timezone-field==7.1
django-two-factor-auth==1.18.1
et_xmlfile==2.0.0
fake-useragent==2.2.0
firebase_admin==7.1.0
fonttools==4.61.1
frozenlist==1.8.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-genai==1.57.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
http_ece==1.2.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
kombu==5.5.4
lxml==6.0.2
msgpack==1.1.2
multidict==6.7.0
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
PyJWT==2.10.1
PyMuPDF==1.26.4
pyOpenSSL==25.3.0
pyparsing==3.3.1
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywebpush==2.1.1
qrcode==8.2
redis==5.2.1
requests==2.32.5
rsa==4.9.1
s3transfer==0.15.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sqlparse==0.5.3
tabulate==0.9.0
tenacity==9.1.2
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==67.0
webauthn==2.7.0
webencodings==0.5.1
websockets==15.0.1
wheel==0.46.3
whitenoise==6.11.0
yarl==1.22.0
zopfli==0.4.0
